Mise en places des statistiques avancées

+++ src/main/java/fr/aimcvent/castleclash/api/utils/TopBoard.java
@@ -15,6 +15,6 @@ public interface TopBoard {
     interface Top {
         String name();
 
-        int value();
+        long value();
     }
 }
